Fix support for the fluid typography
This commit makes the text uses fluid typography if only the `clamp()`
function gets supported.

Before, I thought that if the browsers does not support `clamp()` inside
custom property then it will fallback to use the previous `font-size`
property. It turned out that if the value of the custom property is not
recognized, then the browsers will use `initial` or `inherit` as the
fallback of invalid value of the custom property.