Change the format of the resource lists
This commit uses heading level three and paragraph element instead of a
list element. Also, the data is available on a single file to make it
easier to add and remove links.

This commit also removes the date since I do not see any benefit. I am
also lazy updating the date.